refactor(PropertyTypeArray): rename to PropertyTypeIterable, and preserve old class name for BC only
