+++ tests/Metadata/PropertyTypeIterableTest.php
@@ -7,6 +7,7 @@ namespace Tests\Liip\MetadataParser\Metadata;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Liip\MetadataParser\Metadata\PropertyTypeArray;
+use Liip\MetadataParser\Metadata\PropertyTypeIterable;
 use Liip\MetadataParser\Metadata\PropertyTypeClass;
 use Liip\MetadataParser\Metadata\PropertyTypePrimitive;
 use PHPUnit\Framework\TestCase;
@@ -14,13 +15,13 @@ use PHPUnit\Framework\TestCase;
 /**
  * @small
  */
-class PropertyTypeArrayTest extends TestCase
+class PropertyTypeIterableTest extends TestCase
 {
     public function testNestedArrayLeaf(): void
     {
-        $propertyType = new PropertyTypeArray(new PropertyTypeArray(new PropertyTypePrimitive('int', false), false, false), false, false);
+        $propertyType = new PropertyTypeIterable(new PropertyTypeIterable(new PropertyTypePrimitive('int', false), false, false), false, false);
 
-        $this->assertInstanceOf(PropertyTypeArray::class, $propertyType->getSubType());
+        $this->assertInstanceOf(PropertyTypeIterable::class, $propertyType->getSubType());
         $this->assertInstanceOf(PropertyTypePrimitive::class, $propertyType->getSubType()->getSubType());
         $this->assertInstanceOf(PropertyTypePrimitive::class, $propertyType->getLeafType());
     }
@@ -28,11 +29,14 @@ class PropertyTypeArrayTest extends TestCase
     public function testDefaultListIsNotCollection()
     {
         $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeArray($subType, false, false);
+        $list = new PropertyTypeIterable($subType, false, false);
 
         $this->assertFalse($list->isCollection());
     }
 
+    /**
+     * @deprecated This only checks the behaviour of deprecated class {@see PropertyTypeArray}
+     */
     public function testDefaultCollectionClassIsCollectionInterface()
     {
         $subType = new PropertyTypePrimitive('int', false);
@@ -45,22 +49,17 @@ class PropertyTypeArrayTest extends TestCase
     public function testExplicitCollectionClassIsKept()
     {
         $subType = new PropertyTypePrimitive('int', false);
+        $collection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
 
-        // the $isCollection parameter of {@see PropertyTypeArray::__construct} is only kept for BC,
-        // here we ensure that the metadata still says it's a collection even when that parameter isn't set.
-        foreach ([true, false] as $isCollection) {
-            $collection = new PropertyTypeArray($subType, false, false, $isCollection, ArrayCollection::class);
-
-            $this->assertTrue($collection->isCollection());
-            $this->assertNotEquals(Collection::class, $collection->getCollectionClass());
-            $this->assertEquals(ArrayCollection::class, $collection->getCollectionClass());
-        }
+        $this->assertTrue($collection->isCollection());
+        $this->assertNotEquals(Collection::class, $collection->getCollectionClass());
+        $this->assertEquals(ArrayCollection::class, $collection->getCollectionClass());
     }
 
     public function testMergeListWithClassCollection()
     {
         $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeArray($subType, false, false);
+        $list = new PropertyTypeIterable($subType, false, false);
 
         $this->expectException(\UnexpectedValueException::class);
         $list->merge(new PropertyTypeClass(Collection::class, true));
@@ -73,8 +72,8 @@ class PropertyTypeArrayTest extends TestCase
         $defaultCollection = new PropertyTypeArray($subType, false, false, true);
 
         $result = $defaultCollection->merge($typeHintedCollection);
-        $this->assertInstanceOf(PropertyTypeArray::class, $result);
-        /** @var PropertyTypeArray $result */
+        $this->assertInstanceOf(PropertyTypeIterable::class, $result);
+        /** @var PropertyTypeIterable $result */
         $this->assertTrue($result->isCollection());
         $this->assertEquals($defaultCollection->getSubType(), $result->getSubType());
         $this->assertEquals(Collection::class, $result->getCollectionClass());
@@ -84,11 +83,11 @@ class PropertyTypeArrayTest extends TestCase
     {
         $subType = new PropertyTypePrimitive('int', false);
         $typeHintedCollection = new PropertyTypeClass(Collection::class, true);
-        $explicitCollection = new PropertyTypeArray($subType, false, false, true, ArrayCollection::class);
+        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
 
         $result = $explicitCollection->merge($typeHintedCollection);
-        $this->assertInstanceOf(PropertyTypeArray::class, $result);
-        /** @var PropertyTypeArray $result */
+        $this->assertInstanceOf(PropertyTypeIterable::class, $result);
+        /** @var PropertyTypeIterable $result */
         $this->assertTrue($result->isCollection());
         $this->assertEquals($explicitCollection->getSubType(), $result->getSubType());
         $this->assertEquals(ArrayCollection::class, $result->getCollectionClass());
@@ -98,11 +97,11 @@ class PropertyTypeArrayTest extends TestCase
     {
         $subType = new PropertyTypePrimitive('int', false);
         $defaultCollection = new PropertyTypeArray($subType, false, false, true);
-        $explicitCollection = new PropertyTypeArray($subType, false, false, true, ArrayCollection::class);
+        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
 
         $result = $explicitCollection->merge($defaultCollection);
-        $this->assertInstanceOf(PropertyTypeArray::class, $result);
-        /** @var PropertyTypeArray $result */
+        $this->assertInstanceOf(PropertyTypeIterable::class, $result);
+        /** @var PropertyTypeIterable $result */
         $this->assertTrue($result->isCollection());
         $this->assertEquals($explicitCollection->getSubType(), $result->getSubType());
         $this->assertEquals(ArrayCollection::class, $result->getCollectionClass());
@@ -112,11 +111,11 @@ class PropertyTypeArrayTest extends TestCase
     {
         $subType = new PropertyTypePrimitive('int', false);
         $defaultCollection = new PropertyTypeArray($subType, false, false, true);
-        $explicitCollection = new PropertyTypeArray($subType, false, false, true, ArrayCollection::class);
+        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
 
         $result = $defaultCollection->merge($explicitCollection);
-        $this->assertInstanceOf(PropertyTypeArray::class, $result);
-        /** @var PropertyTypeArray $result */
+        $this->assertInstanceOf(PropertyTypeIterable::class, $result);
+        /** @var PropertyTypeIterable $result */
         $this->assertTrue($result->isCollection());
         $this->assertEquals($explicitCollection->getSubType(), $result->getSubType());
         $this->assertEquals(ArrayCollection::class, $result->getCollectionClass());
@@ -126,11 +125,11 @@ class PropertyTypeArrayTest extends TestCase
     {
         $subType = new PropertyTypePrimitive('int', false);
         $typeHintedCollection = new PropertyTypeClass(Collection::class, true);
-        $explicitCollection = new PropertyTypeArray($subType, false, false, true, ArrayCollection::class);
+        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
 
         $result = $typeHintedCollection->merge($explicitCollection);
-        $this->assertInstanceOf(PropertyTypeArray::class, $result);
-        /** @var PropertyTypeArray $result */
+        $this->assertInstanceOf(PropertyTypeIterable::class, $result);
+        /** @var PropertyTypeIterable $result */
         $this->assertTrue($result->isCollection());
         $this->assertEquals($explicitCollection->getSubType(), $result->getSubType());
         $this->assertEquals(ArrayCollection::class, $result->getCollectionClass());
@@ -139,12 +138,12 @@ class PropertyTypeArrayTest extends TestCase
     public function testMergeListAndCollection()
     {
         $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeArray($subType, false, false);
-        $collection = new PropertyTypeArray($subType, false, false, true);
+        $list = new PropertyTypeIterable($subType, false, false);
+        $collection = new PropertyTypeIterable($subType, false, false, Collection::class);
 
         foreach ([$list->merge($collection), $collection->merge($list)] as $result) {
-            $this->assertInstanceOf(PropertyTypeArray::class, $result);
-            /** @var PropertyTypeArray $result */
+            $this->assertInstanceOf(PropertyTypeIterable::class, $result);
+            /** @var PropertyTypeIterable $result */
             $this->assertTrue($result->isCollection());
             $this->assertEquals(Collection::class, $result->getCollectionClass());
             $this->assertEquals($list->getSubType(), $result->getSubType());
@@ -154,12 +153,12 @@ class PropertyTypeArrayTest extends TestCase
     public function testMergeListWithExplicitCollectionClass()
     {
         $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeArray($subType, false, false);
-        $collection = new PropertyTypeArray($subType, false, false, true, ArrayCollection::class);
+        $list = new PropertyTypeIterable($subType, false, false);
+        $collection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
 
         foreach ([$list->merge($collection), $collection->merge($list)] as $result) {
-            $this->assertInstanceOf(PropertyTypeArray::class, $result);
-            /** @var PropertyTypeArray $result */
+            $this->assertInstanceOf(PropertyTypeIterable::class, $result);
+            /** @var PropertyTypeIterable $result */
             $this->assertTrue($result->isCollection());
             $this->assertEquals(ArrayCollection::class, $result->getCollectionClass());
             $this->assertEquals($list->getSubType(), $result->getSubType());
